refactored bots
