+++ financial-templates-lib/clients/ExpiringMultiPartyClient.js
@@ -1,66 +1,26 @@
-const { delay } = require("./delay");
-const { LiquidationStatesEnum } = require("../common/Enums");
+const { delay } = require("../helpers/delay");
+const { LiquidationStatesEnum } = require("../../common/Enums");
 
-// A thick client for getting information about an ExpiringMultiParty.
-// If no updateThreshold is specified then default to updating every 60 seconds.
+// A thick client for getting information about an ExpiringMultiParty. Used to get sponsor information, outstanding
+// positions, undisputed Liquidations, expired liquidations, disputed liquidations. Client is used by both the
+//  liquidator and dispute bots.
 class ExpiringMultiPartyClient {
-  constructor(logger, abi, web3, empAddress, updateThreshold = 60) {
+  constructor(logger, empAbi, web3, empAddress) {
     this.logger = logger;
-    this.updateThreshold = updateThreshold;
-    this.lastUpdateTimestamp;
-
     this.web3 = web3;
 
     // EMP contract
-    this.emp = new web3.eth.Contract(abi, empAddress);
+    this.emp = new web3.eth.Contract(empAbi, empAddress);
     this.empAddress = empAddress;
 
-    // EMP Data structures & values
+    // EMP Data structures & values to enable synchronous returns of the emp state seen by the client.
     this.sponsorAddresses = [];
     this.positions = [];
     this.undisputedLiquidations = [];
     this.expiredLiquidations = [];
     this.disputedLiquidations = [];
     this.collateralRequirement = null;
-
-    // TODO: Ideally, we'd want to subscribe to events here, but subscriptions don't work with Truffle HDWalletProvider.
-    // One possibility is to experiment with WebSocketProvider instead.
   }
-
-  // Calls _update unless it was recently called, as determined by this.updateThreshold.
-  update = async () => {
-    const currentTime = Math.floor(Date.now() / 1000);
-    if (currentTime < this.lastUpdateTimestamp + this.updateThreshold) {
-      this.logger.debug({
-        at: "ExpiringMultiPartyClient",
-        message: "EMP state update skipped",
-        currentTime: currentTime,
-        lastUpdateTimestamp: this.lastUpdateTimestamp,
-        timeRemainingUntilUpdate: this.lastUpdateTimestamp + this.updateThreshold - currentTime
-      });
-      return;
-    } else {
-      await this._update();
-      this.lastUpdateTimestamp = currentTime;
-      this.logger.debug({
-        at: "ExpiringMultiPartyClient",
-        message: "EMP state updated",
-        lastUpdateTimestamp: this.lastUpdateTimestamp
-      });
-    }
-  };
-
-  // Force call of _update, designed to be called by downstream caller that knowingly updated the EMP state.
-  forceUpdate = async () => {
-    const currentTime = Math.floor(Date.now() / 1000);
-    await this._update();
-    this.lastUpdateTimestamp = currentTime;
-    this.logger.debug({
-      at: "ExpiringMultiPartyClient",
-      message: "EMP state force updated",
-      lastUpdateTimestamp: this.lastUpdateTimestamp
-    });
-  };
 
   // Returns an array of { sponsor, numTokens, amountCollateral } for each open position.
   getAllPositions = () => this.positions;
@@ -95,51 +55,7 @@ class ExpiringMultiPartyClient {
   // Returns an array of sponsor addresses.
   getAllSponsors = () => this.sponsorAddresses;
 
-  start = () => {
-    this._poll();
-  };
-
-  _poll = async () => {
-    while (true) {
-      try {
-        await this._update();
-      } catch (error) {
-        this.logger.error({
-          at: "ExpiringMultiPartyClient",
-          message: "client polling error",
-          error: error
-        });
-      }
-      await delay(Number(10_000));
-    }
-  };
-
-  _isUnderCollateralized = (numTokens, amountCollateral, trv) => {
-    const { toBN, toWei } = this.web3.utils;
-    const fixedPointAdjustment = toBN(toWei("1"));
-    // The formula for an undercollateralized position is:
-    // (numTokens * trv) * collateralRequirement > amountCollateral.
-    // Need to adjust by 10**18 twice because each value is represented as a fixed point scaled up by 10**18.
-    return toBN(numTokens)
-      .mul(toBN(trv))
-      .mul(this.collateralRequirement)
-      .gt(
-        toBN(amountCollateral)
-          .mul(fixedPointAdjustment)
-          .mul(fixedPointAdjustment)
-      );
-  };
-
-  _isExpired = async liquidation => {
-    const currentTime = await this.emp.methods.getCurrentTime().call();
-    return Number(liquidation.liquidationTime) + this.liquidationLiveness <= currentTime;
-  };
-
-  _isLiquidationPreDispute = liquidation => {
-    return liquidation.state === LiquidationStatesEnum.PRE_DISPUTE;
-  };
-
-  _update = async () => {
+  update = async () => {
     this.collateralRequirement = this.web3.utils.toBN(
       (await this.emp.methods.collateralRequirement().call()).toString()
     );
@@ -214,6 +130,30 @@ class ExpiringMultiPartyClient {
             ]),
       []
     );
+  };
+  _isUnderCollateralized = (numTokens, amountCollateral, trv) => {
+    const { toBN, toWei } = this.web3.utils;
+    const fixedPointAdjustment = toBN(toWei("1"));
+    // The formula for an undercollateralized position is:
+    // (numTokens * trv) * collateralRequirement > amountCollateral.
+    // Need to adjust by 10**18 twice because each value is represented as a fixed point scaled up by 10**18.
+    return toBN(numTokens)
+      .mul(toBN(trv))
+      .mul(this.collateralRequirement)
+      .gt(
+        toBN(amountCollateral)
+          .mul(fixedPointAdjustment)
+          .mul(fixedPointAdjustment)
+      );
+  };
+
+  _isExpired = async liquidation => {
+    const currentTime = await this.emp.methods.getCurrentTime().call();
+    return Number(liquidation.liquidationTime) + this.liquidationLiveness <= currentTime;
+  };
+
+  _isLiquidationPreDispute = liquidation => {
+    return liquidation.state === LiquidationStatesEnum.PRE_DISPUTE;
   };
 }
