+++ financial-templates-lib/test/helpers/GasEstimator.js
