+++ monitors/test/CRMonitor.js
@@ -7,7 +7,7 @@ const { interfaceName } = require("../../core/utils/Constants.js");
 const { CRMonitor } = require("../CRMonitor");
 
 // Helper client script
-const { ExpiringMultiPartyClient } = require("../../financial-templates-lib/ExpiringMultiPartyClient");
+const { ExpiringMultiPartyClient } = require("../../financial-templates-lib/clients/ExpiringMultiPartyClient");
 
 // Custom winston transport module to monitor winston log outputs
 const { SpyTransport, lastSpyLogIncludes } = require("../../financial-templates-lib/logger/SpyTransport");
@@ -133,13 +133,13 @@ contract("CRMonitor.js", function(accounts) {
 
   it("Winston correctly emits collateralization ratio message", async function() {
     // No messages created if safely above the CR threshold
-    await empClient._update();
+    await empClient.update();
     await crMonitor.checkWalletCrRatio(time => toWei("1"));
     assert.equal(spy.callCount, 0);
 
     // Emits a message if below the CR threshold. At a price of 1.3 only the monitoredTrader should be undercollateralized
     // with a CR of 250 / (100 * 1.3) =1.923 which is below this addresses threshold of 200 and should emit a message.
-    await empClient._update();
+    await empClient.update();
     await crMonitor.checkWalletCrRatio(time => toWei("1.3"));
     assert.equal(spy.callCount, 1);
     assert.isTrue(lastSpyLogIncludes(spy, "Collateralization ratio alert"));
@@ -150,25 +150,25 @@ contract("CRMonitor.js", function(accounts) {
     // The message should be sent every time the bot is polled and there is a crossing of the threshold line. At a price
     // of 1.2 monitoredTrader's CR = 250/(100*1.2) = 2.083 and monitoredSponsor's CR = 300/(100*1.2) = 2.5 which places
     // both monitored wallets above their thresholds. As a result no new message should be sent.
-    await empClient._update();
+    await empClient.update();
     await crMonitor.checkWalletCrRatio(time => toWei("1.2"));
     assert.equal(spy.callCount, 1); // no new message.
 
     // Crossing the price threshold for both sponsors should emit exactly 2 new messages. At a price of 2.1
     // monitoredTrader's CR = 250/(100*2.1) = 1.1904 and monitoredSponsor's CR = 300/(100*2.1) = 1.42857. At these CRs
     // Both bots are below their thresholds
-    await empClient._update();
+    await empClient.update();
     await crMonitor.checkWalletCrRatio(time => toWei("2.1"));
     assert.equal(spy.callCount, 3); // two new messages
 
     // A second check below this threshold should again trigger messages for both sponsors.
-    await empClient._update();
+    await empClient.update();
     await crMonitor.checkWalletCrRatio(time => toWei("2.1"));
     assert.equal(spy.callCount, 5);
 
     // Reset the price to over collateralized state for both accounts by moving the price into the lower value. This
     // should not emit any events as both correctly collateralized.
-    await empClient._update();
+    await empClient.update();
     await crMonitor.checkWalletCrRatio(time => toWei("1"));
     assert.equal(spy.callCount, 5);
 
@@ -185,7 +185,7 @@ contract("CRMonitor.js", function(accounts) {
 
     await emp.withdrawPassedRequest({ from: monitoredTrader });
 
-    await empClient._update();
+    await empClient.update();
     await crMonitor.checkWalletCrRatio(time => toWei("1"));
     assert.equal(spy.callCount, 6); // a new message is sent.
   });
